requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
shopify-graphql>=0.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.2.0
